% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea-package.R
\docType{package}
\name{ea-package}
\alias{ea}
\alias{ea-package}
\title{ea: Retrieve hydrological data from the Environment Agency}
\description{
Functions to discover and download data from the Environment Agency (EA) Hydrology API.
}
\examples{
\dontrun{
print("Hello, world")
}
}
\author{
\strong{Maintainer}: Simon Moulds \email{simon.moulds@gmail.com} (\href{https://orcid.org/0000-0002-7297-482X}{ORCID})

}
