% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_timeseries_list.R
\name{ea_timeseries_list}
\alias{ea_timeseries_list}
\title{Get available timeseries for one or more stations}
\usage{
ea_timeseries_list(station_guid, ...)
}
\arguments{
\item{station_guid}{Character. Station global unique identifier.
Either a single station ID or a vector of station IDs. This
corresponds to the \code{stationGuid} column in metadata returned
by \code{ea_station_list}.}

\item{...}{Additional arguments. None implemented.}
}
\value{
A tibble.
}
\description{
Get available timeseries for one or more stations
}
\examples{
\dontrun{
# Get discharge stations on the River Thames within 20km of Abingdon
abingdon_stations <- ea_station_list(
  sample_of = "River Thames", observed_property = "waterFlow",
  lat = 51.6708, long = -1.2880, dist = 20
)
guid <- abingdon_stations$stationGuid
tslist <- ea_timeseries_list(guid)
}
}
