% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea-package.R
\docType{package}
\name{ea-package}
\alias{ea}
\alias{ea-package}
\title{ea: Hydrological Data from the Environment Agency}
\description{
Functions to discover and download data from the Environment Agency (EA)
Hydrology API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://simonmoulds.github.io/ea/}
}

}
\author{
\strong{Maintainer}: Simon Moulds \email{sim.moulds@gmail.com} (\href{https://orcid.org/0000-0002-7297-482X}{ORCID})

}
