% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_station_list.R
\name{ea_station_list}
\alias{ea_station_list}
\title{Get tibble with station information}
\usage{
ea_station_list(
  station_id,
  sample_of,
  status_label,
  observed_property,
  open_from,
  open_to,
  long,
  lat,
  easting,
  northing,
  dist,
  ...
)
}
\arguments{
\item{station_id}{Character.}

\item{sample_of}{Character.}

\item{status_label}{Character. One of 'Active', 'Suspended' or}

\item{observed_property}{Character.}

\item{open_from}{Date or character formatted "YYYY-MM-DD"}

\item{open_to}{Date or character formatted "YYYY-MM-DD"}

\item{long}{Numeric. Longitude.}

\item{lat}{Numeric. Latitude.}

\item{easting}{Numeric. Easting.}

\item{northing}{Numeric. Northing.}

\item{dist}{Numeric. Distance in kilometres from the point
specified by \code{long} and \code{lat} or \code{easting} and \code{northing}.}

\item{...}{Additional arguments. None implemented.}
}
\value{
A tibble.
}
\description{
Returns metadata about stations in the EA network.
}
