% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_station_list.R
\name{ea_timeseries_list}
\alias{ea_timeseries_list}
\title{Get list of available time series for one or more stations}
\usage{
ea_timeseries_list(station_id, ...)
}
\arguments{
\item{station_id}{Character or numeric. Either a single station ID or a
vector of station IDs. This corresponds to the \code{notation} column in
metadata returned by \code{ea_station_list}.}

\item{...}{Additional arguments. None implemented.}
}
\value{
A tibble.
}
\description{
Get list of available time series for one or more stations
}
